Allow installing and running the wiki on a server where ini_set has been disabled via disable_functions in php.ini (bug #928168)
The PHPTal skin doesn't function unless the PEAR bits are in include_path,
so it's disabled if ini_set can't be used. If the files are manually placed
or a real PEAR installation is available with the right stuff, one can
use $wgUsePHPTal = true;

There are likely other things that need to be taken into account for
restricted environments. Maintenance scripts haven't been tweaked for
such an environment yet either.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@3519 dd0e9695-b195-4be7-bd10-2dea1a65a6b6