Only show last N contributors in the credits. Unfortunately, this required a change in the getContributors() method of Article, which in turn required some code changes in the Metadata module.
However, it now seems to work.

If the number of contributors is greater than N, then show a link to the
full credits page.

An additional bit: it may be that we don't want to show _any_ contributors
if we are over the max. So, there's a flag to check this.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@4189 dd0e9695-b195-4be7-bd10-2dea1a65a6b6