A major reworking of [[Special:Allmessages]].
So far, the index page has generated 1020 (one thousand and twenty!) SQL
queries. Let's limit this to a maximum of 51. This means that the user is
presented with only 50 rows at a time; however, I have provided "previous" and
"next" links so they can navigate through it.  Additionally, the user can now
enter a "from" term in a text box rather than having to fiddle with the URL.

Furthermore, the user can now choose a namespace using a drop down box.  This
works both on the index page and on a "chunk" page.

Lastly, I think I have made the code for the index page more readable -- not
only by adding comments.

HINT: since this is such a major reworking, the patch is pretty useless.  It is
probably better for anyone interested to view the new file.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@4726 dd0e9695-b195-4be7-bd10-2dea1a65a6b6