* (bug 8641) Fix order of updates to ipblocks table
The update for ipb_anon_only does a total replacement of the ipblocks table in order to apply a unique index at the same time. A later change, adding the ipb_enable_autoblock field, was erroneously listed _before_ this change in the updates list. As a result, upgrades from 1.7 or earlier would end up missing this field as the ipb_anon_only update's new copy of the table dropped the field. A second updater run would restore it, but this should no longer be necessary.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@19382 dd0e9695-b195-4be7-bd10-2dea1a65a6b6