* Don't die when single-element arrays are passed to SQL query constructors   that have an array index other than 0
Since we don't enforce linear ordering of array keys on longer arrays, we
may as well not enforce it on single-element arrays either.
The freaky little Categoryfinder class was using different keys for its
arrays, which could cause a PHP notice and silent lookup failure when
only one ID was looked up.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@32452 dd0e9695-b195-4be7-bd10-2dea1a65a6b6