Re-commit cleaned-up version of r33565 (move subpages along with page). * Move subpages checkbox is not shown unless the current page or its talk page actually has subpages. * Move subpages checkbox is shown if subpages are disabled for the current namespace, provided that the talk page has subpages (but then it has no effect unless the "move talk" box is also checked). * Unnecessary ampersands expunged, superfluous ternary operators excised. * Crazy 1=0 query removed. * Limit of maximum number of pages to move at once added (default 1000).
One niggling annoyance is that if the maximum number of pages is hit due to the base page having too many subpages, the talk page probably won't be moved due to the order that results are returned, which is kind of unexpected.  Also, make sure the queries don't cause MySQL 4 to explode -- MySQL sometimes has problems with OR, especially older versions.  Other than that, should be production-ready, I think.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@35213 dd0e9695-b195-4be7-bd10-2dea1a65a6b6