(bug 5506) Make Title::isAlwaysKnown() return true for known special pages and files in foreign repos.
Besides fixing the display of the latter in logs, edit summaries and other places relying on Linker::link(),
the change should allow considerable simplification (not included in this commit) of that function, as 
well as of several related functions e.g. in the Parser and Skin classes.  To that end, I've also introduced
a new function Title::isKnown() to wrap the common condition "$title->exists() || $title->isAlwaysKnown()",
and have clarified the roles of the three methods.


git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@44520 dd0e9695-b195-4be7-bd10-2dea1a65a6b6