Pull back r44972 for now "(Bug 16677) Media: view URLs should HTTP-redirect to canonical File: namespace"
r44972 isn't quite behaving transparently; non-view URLs are also being redirected to the raw view page, so for instance an ?action=edit or diff or whatever gets lost.

See for example:
http://en.wikipedia.org/wiki/Image:Wiki.png?action=edit
http://en.wikipedia.org/wiki/Media:Wiki.png?action=edit

these both currently behave just like:
http://en.wikipedia.org/wiki/File:Wiki.png?action=edit

while with this patch the Media: one would unexpectedly redirect to the view URL.

The redirection should probably just be done where other existing normalization is... perhaps all that's actually needed is just to replace a Media: Title with the equivalent File: one when initializing the title in the first place, and the existing normalization-redirection will take care of it.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@44981 dd0e9695-b195-4be7-bd10-2dea1a65a6b6