Make "did you mean" more Googley
It wasn't prominent enough; the red text and bolding should make it more
visible, and also more familiar (since I took the styling straight from
Google).  I also moved it up way to the top: it *needs* to be one of the first
things the user sees to be really useful.  This was particularly terrible on
enwiki, where the enormous bloated help message means you have to scroll down
through half a page of largely useless nonsense before even seeing the
suggestion.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@45305 dd0e9695-b195-4be7-bd10-2dea1a65a6b6