When $wgEnforceHtmlIds=false, add legacy anchors
This means that old links will still work in a lot of cases.  However,
if the legacy anchor is invalid XML, we omit it.  In particular, on
non-Latin wikis, practically all old section links (from external
sources or using external links) are likely to break, since the first
character of legacy anchors starting with a non-ASCII character will be
".", which is invalid in XML as well as in HTML4.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@45418 dd0e9695-b195-4be7-bd10-2dea1a65a6b6