Re-implementation of r46725 (caching of Cite <references /> output).
This time, I've written a Parser method called serialiseHalfParsedText, which, as the name implies, grabs some half-parsed text, and fixes up all of the strip
markers, and link comments, and makes it safe to import some other time with unserialiseHalfParsedText. I tested it by live-hacking the cache key to be a constant,
and then putting <references /> on a completely different page, where it worked perfectly.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@46756 dd0e9695-b195-4be7-bd10-2dea1a65a6b6