Merge IEFixes.js into wikibits.js
Not actually tested, but it should theoretically work fine.  This
previously wasn't getting loaded for skins other than Monobook, Modern,
and Vector, and each one had separate code to load it.

I also removed <meta http-equiv="imagetoolbar" content="no" />.  This
was added by Gabriel Wicke without explanation as part of the original
davinci.pt back in r2786, in March 2004.  As far as I can tell, it
disables some minor feature of IE6 for some reason.  I assume it can be
removed completely at this point.  If not, it should be readded without
the conditional comment, which is probably unnecessary.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@60994 dd0e9695-b195-4be7-bd10-2dea1a65a6b6