Provisionally reverting r80430 (bug 529, 12974)
As noted in CR there are some issues where we're trading one set of weird behavior for a different, but previously unknown, set of weird behavior which breaks existing markup that works around the old behavior.
I'd recommend keeping this in store for after the 1.17 stuff calms down so unexpected parser changes aren't cropping up in the middle of things for people working with trunk.

If these are the right changes to make then great -- but they should be done after the consequences are better understood and folks can prepare for changes.




git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@81012 dd0e9695-b195-4be7-bd10-2dea1a65a6b6