Don't regenerate raw sortkeys in CategoryPage.php
Previously, when adding pages, we'd use $title->getCategorySortkey(
$row->cl_sortkey_prefix ) to generate the initial letter on category
pages.  This is fine if the migration script has already been run, but
until then, cl_sortkey_prefix will always be the empty string, so the
user-specified sortkey will be ignored when generating initial letters.
Plus, if getCategorySortkey() had nontrivial computational cost, this
would be a waste of resources.

So just use $row->cl_sortkey instead.  This is at least guaranteed to be
consistent with how pagination works, no matter what state the DB is in.

Already deployed by Roan to fix user bug reports in #wikimedia-tech,
from apergos, Alpha_Quadrant, and Romaine.  Tested on Wikimedia, so far
so good.

git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@83544 dd0e9695-b195-4be7-bd10-2dea1a65a6b6